init vite
