--- app.js
@@ -1,7 +1,0 @@
-import App from "./src/App";
-
-const app = new App({
-  target: document.body,
-});
-
-export default app;
