+++ package.json
@@ -1,11 +1,10 @@
 {
 	"name": "figma2html",
 	"description": "Export Figma frames as images with text elements rendered as HTML",
-	"scripts": {
-		"build": "rollup -c",
-		"dev": "rollup -c -w",
-		"start": "sirv build"
-	},
+	"license": "MIT",
+	"private": true,
+	"main": "dist/main.js",
+	"type": "module",
 	"author": {
 		"name": "Sam Vickars",
 		"email": "sam.vickars@gmail.com",
@@ -21,19 +20,18 @@
 			"name": "Michael Hester"
 		}
 	],
-	"license": "MIT",
+	"scripts": {
+		"dev": "vite",
+		"build": "vite build",
+		"preview": "vite preview",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"postbuild": "tsc -p tsconfig.json"
+	},
 	"devDependencies": {
 		"@figma/plugin-typings": "1.57.1",
-		"@rollup/plugin-alias": "^4.0.2",
-		"@rollup/plugin-babel": "^6.0.3",
-		"@rollup/plugin-commonjs": "24.0.0",
 		"@rollup/plugin-dsv": "^3.0.1",
-		"@rollup/plugin-html": "^1.0.1",
-		"@rollup/plugin-image": "3.0.1",
-		"@rollup/plugin-json": "^6.0.0",
-		"@rollup/plugin-node-resolve": "15.0.1",
-		"@rollup/plugin-terser": "^0.2.1",
-		"@rollup/plugin-typescript": "10.0.1",
+		"@sveltejs/vite-plugin-svelte": "^2.0.2",
+		"@tsconfig/svelte": "^3.0.0",
 		"autoprefixer": "^10.4.13",
 		"caniuse-lite": "1.0.30001441",
 		"cssnano": "^5.1.14",
@@ -46,18 +44,15 @@
 		"postcss-load-config": "^4.0.1",
 		"prettier": "^2.8.1",
 		"prettier-plugin-svelte": "^2.9.0",
-		"rollup": "3.8.1",
-		"rollup-plugin-livereload": "2.0.5",
-		"rollup-plugin-postcss": "^4.0.2",
-		"rollup-plugin-svelte": "7.1.0",
-		"rollup-plugin-svg": "2.0.0",
-		"sirv-cli": "2.0.2",
 		"slugify": "^1.6.5",
 		"svelte": "3.55.0",
+		"svelte-check": "^3.0.1",
 		"svelte-preprocess": "5.0.0",
 		"tailwindcss": "^3.2.4",
 		"tslib": "2.4.1",
-		"typescript": "4.9.4"
+		"typescript": "4.9.4",
+		"vite": "^4.0.4",
+		"vite-plugin-singlefile": "^0.13.2"
 	},
 	"dependencies": {
 		"js-convert-case": "^4.2.0",
