+++ svelte.config.js
@@ -1,0 +1,7 @@
+import sveltePreprocess from 'svelte-preprocess';
+
+export default {
+	preprocess: sveltePreprocess({
+		postcss: true
+	})
+};
