+++ tsconfig.json
@@ -1,8 +1,14 @@
 {
+	"extends": "@tsconfig/svelte/tsconfig.json",
 	"compilerOptions": {
-		"outDir": "build",
-		"target": "ES2019",
+		"outDir": "dist",
+		"target": "ES6",
+		"module": "ES6",
+		"forceConsistentCasingInFileNames": true,
+		"allowJs": true,
+		"checkJs": false,
 		"moduleResolution": "node",
+		"isolatedModules": false,
 		"lib": ["ES2021", "dom"],
 		"typeRoots": [
 			"./node_modules/@types",
@@ -14,6 +20,9 @@
 			"*": ["./src/*"]
 		},
 		"allowSyntheticDefaultImports": true,
-		"resolveJsonModule": true
-	}
+		"resolveJsonModule": true,
+		"sourceMap": false
+	},
+	"include": ["src/**/*.d.ts", "src/**/*.ts", "src/**/*.js", "src/**/*.svelte"],
+	"references": [{ "path": "./tsconfig.node.json" }]
 }
